==> 1_script_scraping.ipynb <==
## Web-Scraping - Game of Thrones - Episode Scripts Of All Seasons

**Quick Notes**

> This is my very first time doing web-scraping. Possible improvement by using an action manager of some sort. Didn't want to waste too much time on this.

This notebook gets all episodes-scripts of the popular HBO series 'Game of Thrones' and writes them into text files. 

The script.txt's are then organized into it's respective season directory which are also created by this file. 

**Sources**

- [Source](https://genius.com/artists/Game-of-thrones)

- [Selenium Documenation](https://selenium-python.readthedocs.io/)

### Import the necessary tools

import time
from selenium import webdriver
from selenium.webdriver.common.by import By
from webdriver_manager.chrome import ChromeDriverManager

### Navigate to get all seasons

# webpage of interest
page_url = "https://genius.com/artists/Game-of-thrones"

# Create driver
driver = webdriver.Chrome(ChromeDriverManager().install())

# Open page url
driver.get(page_url)

# Wait for the cookie banner to load
time.sleep(3)

# Accept cookies
driver.find_element(By.XPATH, '//button[text()="I Accept"]').click()

# Open season overview
driver.find_element(By.XPATH, "/html/body/routable-page/ng-outlet/routable-profile-page/ng-outlet/routed-page/profile-page/div[3]/div[2]/artist-songs-and-albums/album-grid/div[2]").click()

# wait 2 seconds for the script view to load
time.sleep(2)

# create a list with all season elements
season_elements = driver.find_elements(By.CSS_SELECTOR, "[class*='mini_card mini_card--small']")

# Create list with all seasons
seasons = []
for i in range(0, len(season_elements)):
    url = season_elements[i].get_attribute('href')
    desc = season_elements[i].text.replace('Scripts', '')
    seasons.append({'Description': desc, "url": url})

# Check seasons
seasons

### Create Season-Episode catalogues

# Declaring each season as a variable helps me prevent loading time errors when getting each episode...
season_1 = seasons[7]
season_2 = seasons[6]
season_3 = seasons[5]
season_4 = seasons[4]
season_5 = seasons[3]
season_6 = seasons[2]
season_7 = seasons[1]
season_8 = seasons[0]

# get season 1 episodes
driver.get(season_1['url'])

episode_elements = driver.find_elements(By.CSS_SELECTOR, "[class*='u-display_block']")

s1_episodes = []
for ep in episode_elements:
    if ep.get_attribute('href'):
        url = ep.get_attribute('href')
        desc = ep.text.replace(' Lyrics', '')
        s1_episodes.append({'Description': desc, "url": url})

# Check s1 episodes
s1_episodes

# get season 2 episodes
driver.get(season_2['url'])

episode_elements = driver.find_elements(By.CSS_SELECTOR, "[class*='u-display_block']")

s2_episodes = []
for ep in episode_elements:
    if ep.get_attribute('href'):
        url = ep.get_attribute('href')
        desc = ep.text.replace(' Lyrics', '')
        s2_episodes.append({'Description': desc, "url": url})

# Check s2 episodes
s2_episodes

# get season 3 episodes
driver.get(season_3['url'])

episode_elements = driver.find_elements(By.CSS_SELECTOR, "[class*='u-display_block']")

s3_episodes = []
for ep in episode_elements:
    if ep.get_attribute('href'):
        url = ep.get_attribute('href')
        desc = ep.text.replace(' Lyrics', '')
        s3_episodes.append({'Description': desc, "url": url})

# Check s3 episodes
s3_episodes

# get season 4 episodes
driver.get(season_4['url'])

episode_elements = driver.find_elements(By.CSS_SELECTOR, "[class*='u-display_block']")

s4_episodes = []
for ep in episode_elements:
    if ep.get_attribute('href'):
        url = ep.get_attribute('href')
        desc = ep.text.replace(' Lyrics', '')
        s4_episodes.append({'Description': desc, "url": url})

# Check s4 episodes
s4_episodes

# get season 5 episodes
driver.get(season_5['url'])

episode_elements = driver.find_elements(By.CSS_SELECTOR, "[class*='u-display_block']")

s5_episodes = []
for ep in episode_elements:
    if ep.get_attribute('href'):
        url = ep.get_attribute('href')
        desc = ep.text.replace(' Lyrics', '')
        s5_episodes.append({'Description': desc, "url": url})

# Check s5 episodes
s5_episodes

# get season 6 episodes
driver.get(season_6['url'])

episode_elements = driver.find_elements(By.CSS_SELECTOR, "[class*='u-display_block']")

s6_episodes = []
for ep in episode_elements:
    if ep.get_attribute('href'):
        url = ep.get_attribute('href')
        desc = ep.text.replace(' Lyrics', '')
        s6_episodes.append({'Description': desc, "url": url})

# Check s6 episodes
s6_episodes

# get season 7 episodes
driver.get(season_7['url'])

episode_elements = driver.find_elements(By.CSS_SELECTOR, "[class*='u-display_block']")

s7_episodes = []
for ep in episode_elements:
    if ep.get_attribute('href'):
        url = ep.get_attribute('href')
        desc = ep.text.replace(' Lyrics', '')
        s7_episodes.append({'Description': desc, "url": url})

# Check s7 episodes
s7_episodes

# get season 8 episodes
driver.get(season_8['url'])

episode_elements = driver.find_elements(By.CSS_SELECTOR, "[class*='u-display_block']")

s8_episodes = []
for ep in episode_elements:
    if ep.get_attribute('href'):
        url = ep.get_attribute('href')
        desc = ep.text.replace(' Lyrics', '')
        s8_episodes.append({'Description': desc, "url": url})

# Check s8 episodes
s8_episodes

### Create episode script text files and organize them in sub-directorys by season

import os
import shutil

# Function to create sub-dir, create text files and move to subdir
def create_script(fol, season_episodes):
    # if not exists create directory
    if not os.path.exists(fol):
        os.makedirs(fol)
    # get cwd
    cwd = os.getcwd()
    # get fol path
    fol_path = cwd + '/' + str(fol)
    # get season episodes scripts
    for episode in season_episodes:
        # get episode script
        driver.get(episode['url'])
        # Wait for url to load
        time.sleep(3)
        # find script containers
        scripts = driver.find_elements(By.CSS_SELECTOR, "[class*='Lyrics__Container-sc-1ynbvzw-6 jYfhrf'")
        for script in scripts:
            # get file name
            file_name = str(episode['Description']) + '.txt'
            # If file is composed of multiple data containers
            fol_file_path = str(fol_path) + '/' + str(file_name)
            # prints everytime a file is being written (on)
            print(fol_file_path)
            if os.path.isfile(fol_file_path):
                with open(fol_file_path, 'a') as f:
                    f.write('\n')
                    f.write(script.text)
            else:
                # create file
                f = open(file_name,"w+")
                # write in data
                f.write(script.text)
                # wait for write operation
                time.sleep(3)
                # get file path
                file_path = cwd + '/' + str(file_name)
                # move file to season directory
                shutil.move(file_path, fol_path)
        # wait for connection to load
        time.sleep(3)

#### Episodes of Season 1

create_script("Season 1", s1_episodes)

#### Episodes of Season 2

create_script("Season 2", s2_episodes)

#### Episodes of Season 3

create_script("Season 3", s3_episodes)

#### Episodes of Season 4

create_script("Season 4", s4_episodes)

#### Episodes of Season 5

create_script("Season 5", s5_episodes)

#### Episodes of Season 6

create_script("Season 6", s6_episodes)

#### Episodes of Season 7

create_script("Season 7", s7_episodes)

#### Episodes of Season 8

create_script("Season 8", s8_episodes)
