==> 2_character_scraping.ipynb <==
## Web-Scraping - Game of Thrones - Main & Recurring Characters Of All Seasons

**Quick Notes**

This Notebook gets all starring cast characters of the popular HBO series 'Game of Thrones' and organizes them into a '.csv'.

Pandas is truely wonderful.

**Sources**

- [Pandas Documentation](https://pandas.pydata.org/docs/)

### Import the necessary tools

import pandas as pd

### Source

page_url = 'https://en.wikipedia.org/wiki/List_of_Game_of_Thrones_characters'

# Don't need to verify certificate for wikipedia...
import ssl
ssl._create_default_https_context = ssl._create_unverified_context

### Fetch tables from url

table_MN = pd.read_html(page_url)
len(table_MN)

### Identify tables of interest

#### Main characters

main_cast = table_MN[3]
main_cast.head()

main_cast.shape

main_char = main_cast['Character']

main_char.head()

#### Recurring characters

recurr_cast = table_MN[4]
recurr_cast.head()

recurr_cast.shape

recurr_char = recurr_cast['Character']

recurr_char.head()

#### Merge 2 character tables

characters = pd.concat([main_char, recurr_char], ignore_index=True)

characters.shape

characters

import re
# Helper to get nickname
def first_val(list):
    if len(list) > 0:
        return list[0]
    else:
        return 'NaN'
# Extract nick name
characters['Character_nickname'] = characters['Character'].apply(lambda x: re.findall('\".*?\"',x))
characters['Character_nickname'] = characters['Character_nickname'].apply(lambda x: first_val(x).replace('\"', ''))
# Extract first name
characters['Character_firstname'] = characters['Character'].apply(lambda x: x.split(' ', 1)[0])
# Filter from Character name
characters['Character'] = characters['Character'].apply(lambda x: re.sub('\".*?\"','',x))
# Fix double 'space'
characters['Character'] = characters['Character'].apply(lambda x: re.sub('  ',' ', x))

# Check char list(or open csv once created...)
pd.set_option('display.max_rows', None)
characters.head(10)

# Quick fixes char (Names that can be missinterpreted if unchanged)
characters.Character_firstname[17] = 'Drogo' # Khal Drogo
characters.Character_firstname[42] = 'High Sparrow'
characters.Character[42] = 'High Sparrow'
characters.Character_firstname[43] = 'Grey Worm'
characters.Character_firstname[44] = 'Maester Luwin' 
characters.Character_firstname[61] = 'Maester Pycelle'
characters.Character_firstname[63] = 'Maester Aemon'
characters.Character_firstname[98] = 'Black Lorren'
characters.Character_firstname[124] = 'High Septon'
characters.Character_firstname[114] = 'Little Sam'
characters.Character_firstname[132] = 'The Waif' 
characters.Character_firstname[139] = 'Moro' # Khal Moro
characters.Character_firstname[146] = 'Lady Crane'
characters.Character_firstname[149] = 'Black Walder'
characters.Character_firstname[156] = 'Maester Wolkan'
characters.Character_firstname[160] = 'Night King'

### Create 'characters.csv'

string = characters.to_csv(sep=',')

f = open('characters.csv',"w+")
f.write(string)
